RMTP-104 Fix MethodInstantiationInExpressionTreesTest
Due to changed implementation of Enumerable.Empty<T>(),
the returned instance's type has changed from an array to a
custom type. Since the actual return type is not relevant but the
test uses generics (intentionally or by co-incidence) the easiest
way to resolve this is by supplying a generic method that will
be closed based on the supplied proxy type.